Move schema tests and remove code testing behavior of altered Model classes
