+++ tests/postgres_tests/__init__.py
@@ -3,9 +3,7 @@ import unittest
 from forms_tests.widget_tests.base import WidgetTest
 
 from django.db import connection
-from django.test import (
-    SimpleTestCase, TestCase, TransactionTestCase, modify_settings,
-)
+from django.test import SimpleTestCase, TestCase, modify_settings
 
 
 @unittest.skipUnless(connection.vendor == 'postgresql', "PostgreSQL specific tests")
@@ -15,11 +13,6 @@ class PostgreSQLSimpleTestCase(SimpleTestCase):
 
 @unittest.skipUnless(connection.vendor == 'postgresql', "PostgreSQL specific tests")
 class PostgreSQLTestCase(TestCase):
-    pass
-
-
-@unittest.skipUnless(connection.vendor == 'postgresql', "PostgreSQL specific tests")
-class PostgreSQLTransactionTestCase(TransactionTestCase):
     pass
